+++ test/fixtures/callerFilepath.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const common = require('../../');
+
+module.exports = depth => common.callerFilepath(depth);