+++ test/utilities.js
@@ -5,6 +5,8 @@ const common = require('..');
 
 const path = require('path');
 
+const callerFilepathFixture = require('./fixtures/callerFilepath');
+
 metatests.test('Check called filename/filepath', test => {
   test.ok(common.callerFilepath().endsWith(path.join('test', 'utilities.js')));
   test.strictSame(common.callerFilename(), 'utilities.js');
@@ -13,6 +15,7 @@ metatests.test('Check called filename/filepath', test => {
 
 metatests.test('Check called filename/filepath parent', test => {
   child(test, 1);
+  child(test, /child/);
   test.end();
 });
 
@@ -22,6 +25,21 @@ function child(test, depth) {
   );
   test.strictSame(common.callerFilename(depth), 'utilities.js');
 }
+
+metatests.test('Check filepath filter all', test => {
+  test.strictSame(common.callerFilepath(/./), '');
+  test.end();
+});
+
+metatests.test('Check filepath with indirection', test => {
+  test.ok(callerFilepathFixture(1).endsWith(path.join('test', 'utilities.js')));
+  test.ok(
+    callerFilepathFixture(/callerFilepath/).endsWith(
+      path.join('test', 'utilities.js')
+    )
+  );
+  test.end();
+});
 
 metatests.testSync('Check captureMaxStack', test => {
   const stack = common.captureMaxStack();