+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - `common.mkdirpPromise()` function.
 - `Iterator#apply()` and `Iterator#chainApply()` to improve iterator
   interaction with chained calls.
+- `captureMaxStack()` utility to get maximum available stack trace.
 
 ### Changed