+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to
 - `Iterator#apply()` and `Iterator#chainApply()` to improve iterator
   interaction with chained calls.
 - `captureMaxStack()` utility to get maximum available stack trace.
+- Table of contents to documentation.
 
 ### Changed