+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to
 - `captureMaxStack()` utility to get maximum available stack trace.
 - Table of contents to documentation.
 - Get random element from array: `sample(array)`.
+- ECMAScript Modules support via importing '@metarhia/common/module'.
 
 ### Changed