+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to
 - `MemoryWritable` internal buffer size is now limited to 8 MB by default.
 - Signature of `callerFilepath()` to allow passing `RegExp` as depth to be used
   for filtering of stack frames.
+- Return value of `cryptoPrefetcher()` now implements the Iterable interface.
 
 ### Fixed