+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to
 - `captureMaxStack()` utility to get maximum available stack trace.
 - Table of contents to documentation.
 - Get random element from array: `sample(array)`.
-- ECMAScript Modules support via importing '@metarhia/common/module'.
+- ECMAScript Modules named exports support.
 - `Iterator#min()`, `Iterator#max()`, and `Iterator#findCompare()` to
   simplify consumption of iterator in common use-cases
   (finding minimum, maximum, or using a custom condition appropriately).